+++ CHANGELOG.md
@@ -2,6 +2,45 @@
 
 <!--- next entry here -->
 
+## 0.7.0
+2020-04-15
+
+### Features
+
+- including unit test cases (0345c229be4c7009c9338a4b7ac25e0cc59799fd)
+- including logic to merge config files (6a38da056736d4c90e1a00fb79663ad9fab92c02)
+- including logic for garbage collection (49bd62f4547c02d74984dee25cdb3eca6810f6d0)
+
+### Fixes
+
+- including Development Report for week Apr1 to Apr9 (36094e58ec0e7a591302bdc6dfd8808072e1d0e2)
+- minor formatting changes (78468673f0f92875b08770aedc4d870a7a1b44e9)
+- fixing import in test file (5cd85b974174557f6d0229376faa356116113fcc)
+
+## 0.7.0
+2020-04-10
+
+### Features
+
+- including unit test cases (0345c229be4c7009c9338a4b7ac25e0cc59799fd)
+
+### Fixes
+
+- including Development Report for week Apr1 to Apr9 (36094e58ec0e7a591302bdc6dfd8808072e1d0e2)
+- minor formatting changes (78468673f0f92875b08770aedc4d870a7a1b44e9)
+
+## 0.6.0
+2020-04-09
+
+### Features
+
+- added release URLs (68149852bae01bf410637247f0089e54fd3f7668)
+- wrapup CI (757cbf761a4d3659ae34cbc4af9d29c03c1613b7)
+
+### Fixes
+
+- CI (da7333dc9798f636b55465d3574d3ff4fbe98d6d)
+
 ## 0.5.0
 2020-04-09